+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -32,6 +32,7 @@ from imio.helpers.cache import get_current_user_id
 from imio.helpers.content import get_vocab
 from imio.helpers.content import uuidsToObjects
 from imio.helpers.content import uuidToObject
+from imio.helpers.workflow import get_leading_transitions
 from natsort import humansorted
 from operator import attrgetter
 from persistent.list import PersistentList
@@ -141,6 +142,7 @@ from zope.schema.vocabulary import SimpleVocabulary
 
 import copy
 import html
+import itertools
 import logging
 import os
 
@@ -177,18 +179,26 @@ ITEM_WF_STATE_ATTRS = [
     'itemGroupsInChargeStates',
     'itemManualSentToOtherMCStates',
     'itemObserversStates',
-    'recordItemHistoryStates']
+    'recordItemHistoryStates',
+    # datagridfields
+    'powerObservers/item_states']
 ITEM_WF_TRANSITION_ATTRS = [
     'transitionsReinitializingDelays',
     'transitionsToConfirm',
-    'mailItemEvents']
-
+    'mailItemEvents',
+    # datagridfields
+    'onTransitionFieldTransforms/transition',
+    'onMeetingTransitionItemActionToExecute/item_action']
 MEETING_WF_STATE_ATTRS = [
     'itemPreferredMeetingStates',
-    'meetingPresentItemWhenNoCurrentMeetingStates']
+    'meetingPresentItemWhenNoCurrentMeetingStates',
+    # datagridfields
+    'powerObservers/meeting_states']
 MEETING_WF_TRANSITION_ATTRS = [
     'transitionsToConfirm',
-    'mailMeetingEvents']
+    'mailMeetingEvents',
+    # datagridfields
+    'onMeetingTransitionItemActionToExecute/meeting_transition']
 
 schema = Schema((
 
@@ -1242,7 +1252,7 @@ schema = Schema((
             description_msgid="on_transition_field_transforms_descr",
             columns={'transition':
                         SelectColumn("On transition field transform transition",
-                                     vocabulary="listEveryItemTransitions",
+                                     vocabulary="listItemTransitions",
                                      col_description="The transition that will trigger the field transform."),
                      'field_name':
                         SelectColumn("On transition field transform field name",
@@ -1271,7 +1281,7 @@ schema = Schema((
             description_msgid="on_meeting_transition_item_action_to_execute_descr",
             columns={'meeting_transition':
                         SelectColumn("On meeting transition item action to execute meeting transition",
-                                     vocabulary="listEveryMeetingTransitions",
+                                     vocabulary="listMeetingTransitions",
                                      col_description="The transition triggered on the meeting."),
                      'item_action':
                         SelectColumn("On meeting transition item action to execute item action",
@@ -1978,7 +1988,7 @@ schema = Schema((
             i18n_domain='PloneMeeting',
         ),
         schemata="advices",
-        vocabulary='listEveryItemTransitions',
+        vocabulary='listItemTransitions',
         default=defValues.transitionsReinitializingDelays,
         enforceVocabulary=True,
         write_permission="PloneMeeting: Write risky config",
@@ -4616,7 +4626,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
     def validate_meetingConfigsToCloneTo(self, values):
         '''Validates the meetingConfigsToCloneTo.'''
-        # first check that we did not defined to rows for the same meetingConfig
+        # first check that we did not define several rows for the same dest cfg
         meetingConfigs = [v['meeting_config'] for v in values
                           if not v.get('orderindex_', None) == 'template_row_marker']
         tool = api.portal.get_tool('portal_plonemeeting')
@@ -4643,6 +4653,72 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
 
+    def _check_wf_used_in_config(self,
+                                 removed_or_disabled_states=[],
+                                 removed_or_disabled_transitions=[]):
+        """Check if given p_states or p_transitions are used in any MeetingConfig fields."""
+        cfg_item_wf_attrs = list(ITEM_WF_STATE_ATTRS) + list(ITEM_WF_TRANSITION_ATTRS)
+        cfg_meeting_wf_attrs = list(MEETING_WF_STATE_ATTRS) + list(MEETING_WF_TRANSITION_ATTRS)
+        for attr in cfg_item_wf_attrs + cfg_meeting_wf_attrs:
+            # if attr contains a "/" it means it is a column of a datagridfield
+            field = self.getField(attr.split("/")[0])
+            # manage case where item state direclty equal value
+            # or value contains item state, like 'suffix_profile_prereviewers'
+            values = field.getAccessor(self)()
+            if "/" in attr:
+                col_name = attr.split("/")[1]
+                values = [row[col_name] for row in values]
+                # manage multivalued columns
+                if values and hasattr(values[0], "__iter__"):
+                    values = itertools.chain.from_iterable(values)
+            crossed_states = [v for v in values
+                              for r in removed_or_disabled_states if r in v]
+            crossed_transitions = [v for v in values
+                                   for r in removed_or_disabled_transitions if r in v]
+            if crossed_states or crossed_transitions:
+                every_crossed = crossed_states + crossed_transitions
+                return translate(
+                    'state_or_transition_can_not_be_removed_in_use_config',
+                    domain='PloneMeeting',
+                    mapping={
+                        'state_or_transition': translate(
+                            # manage values like MeetingItem.proposed
+                            every_crossed[0].split('.')[-1],
+                            domain="plone",
+                            context=self.REQUEST),
+                        'cfg_field_name': translate(
+                            msgid='PloneMeeting_label_{0}'.format(field.getName()),
+                            domain='PloneMeeting',
+                            context=self.REQUEST)},
+                    context=self.REQUEST)
+        # special check for MeetingConfig.meetingConfigsToCloneTo, current removed
+        # transition could used in another cfg
+        cfg_id = self.getId()
+        tool = api.portal.get_tool('portal_plonemeeting')
+        for other_cfg in tool.objectValues('MeetingConfig'):
+            if other_cfg == self:
+                continue
+            values = [v['trigger_workflow_transitions_until'].split('.')[1]
+                      for v in other_cfg.getMeetingConfigsToCloneTo()
+                      if v['meeting_config'] == cfg_id and
+                      v['trigger_workflow_transitions_until'].split('.')[1]
+                      in removed_or_disabled_transitions]
+            if values:
+                return translate(
+                    'state_or_transition_can_not_be_removed_in_use_other_config',
+                    domain='PloneMeeting',
+                    mapping={
+                        'transition': translate(
+                            values[0],
+                            domain="plone",
+                            context=self.REQUEST),
+                        'parameter_label': translate(
+                            "PloneMeeting_label_meetingConfigsToCloneTo",
+                            domain="PloneMeeting",
+                            context=self.REQUEST),
+                        'other_cfg_title': safe_unicode(other_cfg.Title()), },
+                    context=self.REQUEST)
+
     security.declarePrivate('validate_workflowAdaptations')
 
     def validate_workflowAdaptations(self, values):
@@ -4655,18 +4731,83 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         if '' in values:
             values.remove('')
 
+        # used to check removed and conflicts
+        item_type = self.getItemTypeName()
+        meeting_type = self.getMeetingTypeName()
+        validation_returned_states = _getValidationReturnedStates(self)
+        removed_and_conflicts_checks = OrderedDict([
+            ('hide_decisions_when_under_writing',
+             {'portal_type': meeting_type,
+              'review_state': ['decisions_published'],
+              'optional_with': ()}),
+            ('accepted_out_of_meeting',
+             {'portal_type': item_type,
+              'review_state': ['accepted_out_of_meeting'],
+              'optional_with': ('accepted_out_of_meeting_and_duplicated', )}),
+            ('accepted_out_of_meeting_and_duplicated',
+             {'portal_type': item_type,
+              'review_state': ['accepted_out_of_meeting'],
+              'optional_with': ('accepted_out_of_meeting', )}),
+            ('accepted_out_of_meeting_emergency',
+             {'portal_type': item_type,
+              'review_state': ['accepted_out_of_meeting_emergency'],
+              'optional_with': ('accepted_out_of_meeting_emergency_and_duplicated', )}),
+            ('accepted_out_of_meeting_emergency_and_duplicated',
+             {'portal_type': item_type,
+              'review_state': ['accepted_out_of_meeting_emergency'],
+              'optional_with': ('accepted_out_of_meeting_emergency', )}),
+            ('transfered',
+             {'portal_type': item_type,
+              'review_state': ['transfered'],
+              'optional_with': ('transfered_and_duplicated', )}),
+            ('transfered_and_duplicated',
+             {'portal_type': item_type,
+              'review_state': ['transfered'],
+              'optional_with': ('transfered', )}),
+            ('removed',
+             {'portal_type': item_type,
+              'review_state': ['removed'],
+              'optional_with': ('removed_and_duplicated', )}),
+            ('removed_and_duplicated',
+             {'portal_type': item_type,
+              'review_state': ['removed'],
+              'optional_with': ('removed', )}),
+            ('postpone_next_meeting',
+             {'portal_type': item_type,
+              'review_state': ['postponed_next_meeting'],
+              'optional_with': ()}),
+            ('mark_not_applicable',
+             {'portal_type': item_type,
+              'review_state': ['marked_not_applicable'],
+              'optional_with': ()}),
+            ('refused',
+             {'portal_type': item_type,
+              'review_state': ['refused'],
+              'optional_with': ()}),
+            ('delayed',
+             {'portal_type': item_type,
+              'review_state': ['delayed'],
+              'optional_with': ()}),
+            ('accepted_but_modified',
+             {'portal_type': item_type,
+              'review_state': ['accepted_but_modified'],
+              'optional_with': ()}),
+            ('pre_accepted',
+             {'portal_type': item_type,
+              'review_state': ['pre_accepted'],
+              'optional_with': ()}),
+            ('return_to_proposing_group',
+             {'portal_type': item_type,
+              'review_state': ['returned_to_proposing_group'],
+              'optional_with': ('return_to_proposing_group_with_last_validation',
+                                'return_to_proposing_group_with_all_validations')}),
+            ('return_to_proposing_group_with_last_validation',
+             {'portal_type': item_type,
+              'review_state': validation_returned_states,
+              'optional_with': ('return_to_proposing_group_with_all_validations', )})])
+
         # conflicts
         msg = translate('wa_conflicts', domain='PloneMeeting', context=self.REQUEST)
-        if 'removed' in values and 'removed_and_duplicated' in values:
-            return msg
-        if 'accepted_out_of_meeting' in values and \
-           'accepted_out_of_meeting_and_duplicated' in values:
-            return msg
-        if 'accepted_out_of_meeting_emergency' in values and \
-           'accepted_out_of_meeting_emergency_and_duplicated' in values:
-            return msg
-        if 'transfered' in values and 'transfered_and_duplicated' in values:
-            return msg
         if 'no_decide' in values and 'hide_decisions_when_under_writing' in values:
             return msg
         # several 'return_to_proposing_group' values may not be selected together
@@ -4674,6 +4815,12 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             v for v in values if v.startswith('return_to_proposing_group')]
         if len(return_to_prop_group_wf_adaptations) > 1:
             return msg
+        # check removed_checks taking into account the "optional_with" values
+        # that links 2 wfa that can not be used together
+        for conflict_wfa, infos in removed_and_conflicts_checks.items():
+            if infos['optional_with'] and \
+               (conflict_wfa in values and set(infos['optional_with']).intersection(values)):
+                return msg
 
         # dependecies, some adaptations will complete already select ones
         dependencies = {
@@ -4726,7 +4873,8 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
                         context=self.REQUEST)
 
         catalog = api.portal.get_tool('portal_catalog')
-        wfTool = api.portal.get_tool('portal_workflow')
+        itemWF = self.getItemWorkflow(theObject=True)
+        meetingWF = self.getMeetingWorkflow(theObject=True)
 
         # validate new added workflowAdaptations regarding existing items and meetings
         added = set(values).difference(set(self.getWorkflowAdaptations()))
@@ -4734,9 +4882,9 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             # this will remove the 'published' state for Meeting and 'itempublished' for MeetingItem
             # check that no more elements are in these states
             if catalog.unrestrictedSearchResults(
-                portal_type=self.getItemTypeName(), review_state='itempublished') or \
+                portal_type=item_type, review_state='itempublished') or \
                catalog.unrestrictedSearchResults(
-                    portal_type=self.getMeetingTypeName(), review_state='published'):
+                    portal_type=meeting_type, review_state='published'):
                 return translate('wa_added_no_publication_error',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
@@ -4744,9 +4892,9 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             # this will remove the 'frozen' state for Meeting and 'itemfrozen' for MeetingItem
             # check that no more elements are in these states
             if catalog.unrestrictedSearchResults(
-                portal_type=self.getItemTypeName(), review_state='itemfrozen') or \
+                portal_type=item_type, review_state='itemfrozen') or \
                catalog.unrestrictedSearchResults(
-                    portal_type=self.getMeetingTypeName(), review_state='frozen'):
+                    portal_type=meeting_type, review_state='frozen'):
                 return translate('wa_added_no_freeze_error',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
@@ -4754,7 +4902,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             # this will remove the 'decided' state for Meeting
             # check that no more elements are in these states
             if catalog.unrestrictedSearchResults(
-                    portal_type=self.getMeetingTypeName(), review_state='decided'):
+                    portal_type=meeting_type, review_state='decided'):
                 return translate('wa_added_no_decide_error',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
@@ -4770,143 +4918,62 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             # get every 'waiting_advices'-like states, we could have 'itemcreated_waiting_advices',
             # 'proposed_waiting_advices' or
             # 'itemcreated__or__proposedToValidationLevel1__or__..._waiting_advices' for example
-            itemWF = wfTool.getWorkflowsFor(self.getItemTypeName())[0]
             waiting_advices_states = [state for state in itemWF.states if 'waiting_advices' in state]
             if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state=waiting_advices_states):
+                    portal_type=item_type, review_state=waiting_advices_states):
                 return translate('wa_removed_waiting_advices_error',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
-        if ('return_to_proposing_group' in removed) and \
-            not (('return_to_proposing_group_with_last_validation' in added) or
-                 ('return_to_proposing_group_with_all_validations' in added)):
-            # this will remove the 'returned_to_proposing_group' state for MeetingItem
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='returned_to_proposing_group'):
-                return translate('wa_removed_return_to_proposing_group_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        validation_returned_states = _getValidationReturnedStates(self)
-        if ('return_to_proposing_group_with_last_validation' in removed) and \
-           not ('return_to_proposing_group_with_all_validations' in added):
-            # this will remove the 'returned_to_proposing_group with last validation state'
-            # for MeetingItem check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state=validation_returned_states):
-                return translate('wa_removed_return_to_proposing_group_with_last_validation_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
+
         if 'return_to_proposing_group_with_all_validations' in removed:
             # this will remove the 'returned_to_proposing_group with every validation states'
             # for MeetingItem check that no more items are in these states
             # not downgrade from all to last validation if one item is in intermediary state
             if (catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state=validation_returned_states)) or \
+                    portal_type=item_type, review_state=validation_returned_states)) or \
                (('return_to_proposing_group' not in added) and
                 ('return_to_proposing_group_with_last_validation' not in added) and
                     (catalog.unrestrictedSearchResults(
-                        portal_type=self.getItemTypeName(), review_state='returned_to_proposing_group'))):
+                        portal_type=item_type, review_state='returned_to_proposing_group'))):
                 return translate('wa_removed_return_to_proposing_group_with_all_validations_error',
                                  domain='PloneMeeting',
                                  context=self.REQUEST)
-        if 'hide_decisions_when_under_writing' in removed:
-            # this will remove the 'decisions_published' state for Meeting
-            # check that no more meetings are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getMeetingTypeName(),
-                    review_state='decisions_published'):
-                return translate('wa_removed_hide_decisions_when_under_writing_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'accepted_out_of_meeting' in removed or \
-           'accepted_out_of_meeting_and_duplicated' in removed:
-            # this will remove the 'accepted_out_of_meeting' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(),
-                    review_state='accepted_out_of_meeting'):
-                return translate('wa_removed_accepted_out_of_meeting_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'accepted_out_of_meeting_emergency' in removed or \
-           'accepted_out_of_meeting_emergency_and_duplicated' in removed:
-            # this will remove the 'accepted_out_of_meeting_emergency' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(),
-                    review_state='accepted_out_of_meeting_emergency'):
-                return translate('wa_removed_accepted_out_of_meeting_emergency_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'transfered' in removed or 'transfered_and_duplicated' in removed:
-            # this will remove the 'transfered' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(),
-                    review_state='transfered'):
-                return translate('wa_removed_transfered_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'postpone_next_meeting' in removed:
-            # this will remove the 'postponed_next_meeting' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='postponed_next_meeting'):
-                return translate('wa_removed_postpone_next_meeting_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'mark_not_applicable' in removed:
-            # this will remove the 'marked_not_applicable' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='marked_not_applicable'):
-                return translate('wa_removed_mark_not_applicable_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'refused' in removed:
-            # this will remove the 'refused' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='refused'):
-                return translate('wa_removed_refused_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'delayed' in removed:
-            # this will remove the 'delayed' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='delayed'):
-                return translate('wa_removed_delayed_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'accepted_but_modified' in removed:
-            # this will remove the 'accepted_but_modified' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='accepted_but_modified'):
-                return translate('wa_removed_accepted_but_modified_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        if 'pre_accepted' in removed:
-            # this will remove the 'pre_accepted' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='pre_accepted'):
-                return translate('wa_removed_pre_accepted_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
-        # check if 'removed' is removed only if the other 'removed_and_duplicated' is not added
-        # at the same time (switch from one to other)
-        if ('removed' in removed and 'removed_and_duplicated' not in added) or \
-           ('removed_and_duplicated' in removed and 'removed' not in added):
-            # this will remove the 'removed' state for Item
-            # check that no more items are in this state
-            if catalog.unrestrictedSearchResults(
-                    portal_type=self.getItemTypeName(), review_state='removed'):
-                return translate('wa_removed_removed_error',
-                                 domain='PloneMeeting',
-                                 context=self.REQUEST)
+
+        # check removed directly linked to a single review_state
+        for removed_wfa, infos in removed_and_conflicts_checks.items():
+            if removed_wfa in removed and \
+               (not infos['optional_with'] or
+                    not set(infos['optional_with']).intersection(added)):
+                # check that no more elements are in removed state
+                if catalog.unrestrictedSearchResults(
+                        portal_type=infos['portal_type'],
+                        review_state=infos['review_state']):
+                    return translate(
+                        'wa_removed_found_elements_error',
+                        mapping={
+                            'wfa': translate(
+                                "wa_%s" % removed_wfa,
+                                domain="PloneMeeting",
+                                context=self.REQUEST),
+                            'review_state': translate(
+                                infos['review_state'][0],
+                                domain="plone",
+                                context=self.REQUEST)},
+                        domain='PloneMeeting',
+                        context=self.REQUEST)
+                else:
+                    # check that removed states/transitions no more used in config
+                    # we have the states, get the transitions leading to it
+                    related_wf = itemWF if infos['portal_type'] == item_type else meetingWF
+                    removed_or_disabled_transitions = list(itertools.chain.from_iterable(
+                        [[tr.id for tr in get_leading_transitions(related_wf, state_id)]
+                         for state_id in infos['review_state']]))
+                    used_in_cfg_error_msg = self._check_wf_used_in_config(
+                        removed_or_disabled_states=infos['review_state'],
+                        removed_or_disabled_transitions=removed_or_disabled_transitions)
+                    if used_in_cfg_error_msg:
+                        return used_in_cfg_error_msg
+
         return self.adapted().custom_validate_workflowAdaptations(values, added, removed)
 
     def custom_validate_workflowAdaptations(self, values, added, removed):
@@ -5009,9 +5076,6 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
 
         # make sure the MeetingConfig does not use a state that was removed or disabled
         # either using state or transition
-        # states
-        cfg_item_wf_attrs = list(ITEM_WF_STATE_ATTRS) + list(ITEM_WF_TRANSITION_ATTRS)
-        # transitions
         enabled_stored_transitions = self.getItemWFValidationLevels(
             data='leading_transition',
             only_enabled=True)
@@ -5021,35 +5085,9 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
             only_enabled=True)
         removed_or_disabled_transitions = tuple(set(enabled_stored_transitions).difference(
             set(enabled_values_transitions)))
-
-        for attr in cfg_item_wf_attrs:
-            field = self.getField(attr)
-            # manage case where item state direclty equal value
-            # of value contains item state, like 'suffix_profile_prereviewers'
-            crossed_states = [v for v in field.getAccessor(self)()
-                              for r in removed_or_disabled_states if r in v]
-            crossed_transitions = [v for v in field.getAccessor(self)()
-                                   for r in removed_or_disabled_transitions if r in v]
-            if crossed_states or crossed_transitions:
-                every_crossed = crossed_states + crossed_transitions
-                return translate(
-                    'item_wf_val_states_can_not_be_removed_in_use_config',
-                    domain='PloneMeeting',
-                    mapping={
-                        'state_or_transition': translate(
-                            # manage values like MeetingItem.proposed
-                            every_crossed[0].split('.')[-1],
-                            domain="plone",
-                            context=self.REQUEST),
-                        'cfg_field_name': translate(
-                            msgid='PloneMeeting_label_{0}'.format(field.getName()),
-                            domain='PloneMeeting',
-                            context=self.REQUEST)},
-                    context=self.REQUEST)
-
-        # XXX to be managed when moving MeetingConfig to DX
-        # some datagridfield attributes
-        # powerObservers, meetingConfigsToCloneTo, onTransitionFieldTransforms
+        return self._check_wf_used_in_config(
+            removed_or_disabled_states=removed_or_disabled_states,
+            removed_or_disabled_transitions=removed_or_disabled_transitions)
 
     security.declarePrivate('validate_mailItemEvents')
 
@@ -6962,18 +7000,18 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         '''Vocabulary for column item_action of field onMeetingTransitionItemActionToExecute.
            This list a special value 'Execute given action' and the list of item transitions.'''
         res = [(EXECUTE_EXPR_VALUE, _(EXECUTE_EXPR_VALUE))]
-        transitions = self.listEveryItemTransitions()
+        transitions = self.listItemTransitions()
         return DisplayList(res) + transitions
 
-    security.declarePrivate('listEveryItemTransitions')
+    security.declarePrivate('listItemTransitions')
 
-    def listEveryItemTransitions(self):
+    def listItemTransitions(self):
         '''Vocabulary that list every item WF transitions.'''
         return DisplayList(self.listTransitions('Item')).sortedByValue()
 
-    security.declarePrivate('listEveryMeetingTransitions')
+    security.declarePrivate('listMeetingTransitions')
 
-    def listEveryMeetingTransitions(self):
+    def listMeetingTransitions(self):
         '''Vocabulary that list every meeting WF transitions.'''
         return DisplayList(self.listTransitions('Meeting')).sortedByValue()